requirements.txt
jinja2
werkzeug
gunicorn
fabric3

setproctitle
